Aggregation of multiple model changes
This commit has a significant amount of changes to the way the models
work. I don't remember if the previous commit had any model details in
it but I changed it half way through because it made more sense to use
Worn as a parent table. WornLayers and WornPhotos is now a child table.